+++ CHANGELOG.md
@@ -162,7 +162,7 @@
 * Define `TT_T_I` data type
 * Define `U_UU_UUU_V_I_II_III_UUUU` data type
 * Define `Flip` type family
-* Define `Order` type family
+* Define `Flippable` type family
 * Define `absurd` expression
 * Define `empty` expression
 * Define `_lj` operator